+++ example/bases.dart
@@ -21,12 +21,12 @@ void main(List<String> args) {
   print(dec);
   print(dec.checkMatrix);
 
-  final C = Matrix([
-    [0, 1, 1],
-    [2, 1, 0],
-    [3, 4, 5]
+  var matrix = Matrix([
+    [4, 2, 3],
+    [2, 4, 5],
+    [3, 5, 8]
   ]);
-  dec = C.decomposition.eigenvalueDecomposition();
-  print(dec);
-  print(dec.checkMatrix);
+
+  var eigen = matrix.decomposition.choleskyDecomposition();
+  print(eigen);
 }